requirements.txt
django
django-environ
djangorestframework
django-cors-headers
django-filter
django-extensions
psycopg2-binary
django-redis
whitenoise
djangorestframework-simplejwt
argon2-cffi
django-multitenant
requests
gunicorn
celery
redis
